refactor: extract pair ratios calculation to hook
