+++ packages/web/components/home/chart-section.tsx
@@ -1,8 +1,6 @@
-import { TokenHistoricalPrice } from "@osmosis-labs/stores/build/queries-external/token-historical-chart/types";
 import classNames from "classnames";
 import Image from "next/image";
 import React, { useState } from "react";
-import { useMemo } from "react";
 
 import { Icon } from "~/components/assets";
 import { DoubleTokenChart } from "~/components/chart/double-token-chart";
@@ -12,6 +10,7 @@ import {
   useSwapHistoricalPrice,
   useSwapPageQuery,
 } from "~/components/home/hooks";
+import { useCalculatePairRatios } from "~/hooks/use-calculate-pair-ratios";
 import { CommonPriceChartTimeFrame } from "~/server/queries/complex/assets";
 import { theme } from "~/tailwind.config";
 import * as trpc from "~/utils/trpc";
@@ -22,41 +21,6 @@ const availableTimeFrames: CommonPriceChartTimeFrame[] = [
   "1W",
   "1M",
 ];
-
-type AssetChartData =
-  | (TokenHistoricalPrice & {
-      denom: string;
-    })[];
-
-const calculatePairRatios = (
-  from?: AssetChartData,
-  to?: AssetChartData
-): AssetChartData => {
-  const ratios: AssetChartData = [];
-
-  if (from && to && from.length === to.length) {
-    from.forEach((from, i) => {
-      // Calculate ratio for each property
-      const closeRatio = from.close / to[i].close;
-      const highRatio = from.high / to[i].high;
-      const lowRatio = from.low / to[i].low;
-      const openRatio = from.open / to[i].open;
-      const volumeRatio = from.volume / to[i].volume;
-
-      // Push ratios to the array
-      ratios.push({
-        denom: `${from.denom}/${to[i].denom}`,
-        close: closeRatio,
-        high: highRatio,
-        low: lowRatio,
-        open: openRatio,
-        volume: volumeRatio,
-        time: (from.time + to[i].time) / 2,
-      });
-    });
-  }
-  return ratios;
-};
 
 export const ChartSection = ({
   isChartVisible,
@@ -89,9 +53,9 @@ export const ChartSection = ({
   const [showPairRatio, setShowPairRatio] = useState(false);
   const [isTimeFrameSelectorOpen, setIsTimeFrameSelectorOpen] = useState(false);
 
-  const pairRatios = useMemo(
-    () => calculatePairRatios(fromAssetChartData, toAssetChartData),
-    [fromAssetChartData, toAssetChartData]
+  const pairRatios = useCalculatePairRatios(
+    fromAssetChartData,
+    toAssetChartData
   );
 
   return (
